feat: ディレクトリにComponentとContainerを別で作るようにした
