+++ src/organisms/RankingPageTable/index.js
@@ -1,0 +1,3 @@
+import RankingPageTableContainer from "./RankingPageTableContainer";
+
+export default RankingPageTableContainer;
